BUG - not blocking waiting for socket to open
I'm not sure the whole design is correct.  I think the waiting for the
socket etc to be open might be the wrong design.  I think by default
it will only send when the socket is open.  It might be enough to
check if its in the correct state, if not create a new websocket.